bot-client/requirements.txt
aiogram==3.22.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
